Move sockets to individual package
