+++ be1-go/hub/mod.go
@@ -5,11 +5,56 @@
 //
 // A concrete instance of a Hub may be an Organizer or a Witness. A baseHub
 // type contains the implementation common across both types.
-//
-// The package also contains an implmentation of the Socket interface which
-// is responsible for low level communication over websockets, i.e. sending
-// and receiving marshaled messages over the wire.
-//
-// The Socket interface has multiple concrete implmentations - one for each
-// type of client the backend interacts with: an Attendee, Organizer or Witness.
 package hub
+
+import (
+	"context"
+	"student20_pop/message"
+	"student20_pop/network/socket"
+)
+
+// HubType denotes the type of the hub.
+type HubType string
+
+const (
+	// OrganizerHubType represents the Organizer Hub.
+	OrganizerHubType HubType = "organizer"
+
+	// WitnessHubType represnets the Witness Hub.
+	WitnessHubType HubType = "witness"
+)
+
+// Hub defines the methods a PoP server must implement to receive messages
+// and handle clients.
+type Hub interface {
+	// Start invokes the processing loop for the hub.
+	Start(ctx context.Context)
+
+	// Receiver returns a channel that may be used to
+	// process incoming messages
+	Receiver() chan<- socket.IncomingMessage
+
+	// OnSocketClose returns a channel which accepts
+	// socket ids on connection close events. This
+	// allows the hub to cleanup clients which close
+	// without sending an unsubscribe message
+	OnSocketClose() chan<- string
+
+	// Type returns the type of Hub.
+	Type() HubType
+}
+
+// Channel represents a PoP channel - like a LAO.
+type Channel interface {
+	// Subscribe is used to handle a subscribe message.
+	Subscribe(socket socket.Socket, msg message.Subscribe) error
+
+	// Unsubscribe is used to handle an unsubscribe message.
+	Unsubscribe(socketID string, msg message.Unsubscribe) error
+
+	// Publish is used to handle a publish message.
+	Publish(msg message.Publish) error
+
+	// Catchup is used to handle a catchup message.
+	Catchup(msg message.Catchup) []message.Message
+}
